move call to STS move api into project.ts
